fairlens_backend/requirements.txt
fastapi==0.103.2
uvicorn[standard]==0.23.2
python-dotenv==1.0.0
httpx==0.25.2
pydantic==2.4.2
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.4
joblib==1.3.2
scipy==1.11.4
